pyproject.toml
[tool.poetry]
name = "git-repo-info"
version = "0.1.0"
description = "A tool for retrieving information about Git repositories"
authors = ["James Medlin <jmedlin@westervelt.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
rich = "^13.9.4"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-mock = "^3.14.0"
pytest-cov = "^6.0.0"
pre-commit = "^4.0.1"
pre-commit-hooks = "^5.0.0"
isort = "^5.13.2"
autoflake = "^2.3.1"
flake8 = "^7.1.1"
pyupgrade = "^3.19.1"
black = "^24.10.0"
vulture = "^2.14"

[tool.poetry.scripts]
git-repo-info = "git_repo_info:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# This setting should be the same as the pyproject.toml file in /hq-util4
[tool.black]
line-length = 115

# This setting should be the same as the pyproject.toml file in /hq-util4
[tool.isort]
profile = "black"

[tool.pytest.ini_options]
log_file = "tests/test.log"
log_file_level = "DEBUG"
log_file_format = "[%(asctime)s] %(levelname)s: %(message)s"
log_file_date_format = "%H:%M:%S"

[tool.vulture]
paths = ["git_repo_info", "whitelist.py"]
sort_by_size = false
verbose = false
